requirements.txt
anthropic>=0.40.0
rich>=14.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
aiofiles>=23.0.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.0.0
python-telegram-bot>=21.0.0
duckduckgo-search>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
Pillow>=10.0.0
pdfplumber>=0.10.0
pandas>=2.0.0
pygments>=2.18.0
APScheduler>=3.10.0
cryptography>=42.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
mypy>=1.10.0
